fix: correct formulas for given plot
now Y can be fractional

+++ web/1/lib/dto.go
@@ -1,9 +1,9 @@
 package lib
 
 type DotParams struct {
-	X string `json:"X"`
-	Y int    `json:"Y"`
-	R string `json:"R"`
+	X string  `json:"X"`
+	Y float64 `json:"Y"`
+	R string  `json:"R"`
 }
 
 type DotStatus struct {
